Proof of concept RNG from R
