+++ NAMESPACE
@@ -1,3 +1,3 @@
 # Generated by roxygen2: do not edit by hand
 
-useDynLib(dust, .registration = TRUE)
+useDynLib(dust)
