+++ src/test_rng.h
@@ -1,0 +1,5 @@
+#define STRICT_R_HEADERS
+#include <R.h>
+#include <Rinternals.h>
+
+extern "C" SEXP test_rng();
