+++ tests/testthat/test-rng.R
@@ -1,0 +1,3 @@
+context("rng")
+
+.Call("test_rng", PACKAGE = "dust")
